리사이클러 뷰에 특정 지역 안나오는거 나오게 해결 .

+++ main_mango/app/src/main/java/com/softSquared/mangoplate/src/ApplicationClass.java
@@ -28,8 +28,8 @@ public class ApplicationClass extends Application { //쓸꺼면 여기에 상수
     //널이면 재할당.
 
     private static ApplicationClass instance;
-    public static double lat;
-    public static double lng;
+    public static float lat;
+    public static float lng;
     public static ApplicationClass getGlobalApplicationContext() {
         if (instance == null) {
             throw new IllegalStateException("This Application does not inherit com.kakao.GlobalApplication");
